requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
configparser==4.0.2
contextlib2==0.6.0.post1
Django==1.11.26
django-filter==1.1.0
djangorestframework==3.9.4
funcsigs==1.0.2
idna==2.8
importlib-metadata==1.3.0
jsonify==0.5
more-itertools==5.0.0
packaging==19.2
pathlib2==2.3.5
pdfminer2==20151206
pdfx==1.3.0
pluggy==0.13.1
py==1.8.0
pyparsing==2.4.5
PyPDF2==1.26.0
pytest==4.6.7
pytz==2019.3
requests==2.22.0
scandir==1.10.0
six==1.13.0
urllib3==1.25.7
wcwidth==0.1.7
zipp==0.6.0
